feat(draftjs): Move tests
